requirements.txt
torch
torchvision
numpy
pandas
scikit-learn
pillow
tqdm
requests
matplotlib
seaborn
